+++ benchopt/tests/dummy_benchmark/solvers/r_pgd.py
@@ -1,25 +1,20 @@
 from pathlib import Path
 
 from benchopt import BaseSolver
-from benchopt import safe_import_context
 
+import numpy as np
+from scipy.sparse import issparse
 
-with safe_import_context() as import_ctx:
-    import numpy as np
-    from scipy.sparse import issparse
+# Import helpers from rpy2 and benchopt.helpers.r_lang
+from rpy2 import robjects
+from rpy2.robjects import numpy2ri
+from rpy2.robjects import default_converter
+from benchopt.helpers.r_lang import import_func_from_r_file
 
-    # Import helpers from rpy2 and benchopt.helpers.r_lang
-    from rpy2 import robjects
-    from rpy2.robjects import numpy2ri
-    from benchopt.helpers.r_lang import import_func_from_r_file
-
-    # Setup the system to allow passing numpy arrays to rpy2
-    numpy2ri.activate()
-
-    # Import R function defined in r_pgd.R so they can be retrieved as python
-    # functions using `func = robjects.r['FUNC_NAME']`
-    R_FILE = str(Path(__file__).with_suffix('.R'))
-    import_func_from_r_file(R_FILE)
+# Import R function defined in r_pgd.R so they can be retrieved as python
+# functions using `func = robjects.r['FUNC_NAME']`
+R_FILE = str(Path(__file__).with_suffix('.R'))
+import_func_from_r_file(R_FILE)
 
 
 class Solver(BaseSolver):
@@ -39,8 +34,11 @@ class Solver(BaseSolver):
         self.r_pgd = robjects.r['proximal_gradient_descent']
 
     def run(self, n_iter):
-        coefs = self.r_pgd(self.X, self.y[:, None], self.lmbd, n_iter=n_iter)
-        self.w = np.asarray(coefs)
+        with (default_converter + numpy2ri.converter).context():
+            coefs = self.r_pgd(
+                self.X, self.y[:, None], self.lmbd, n_iter=n_iter
+            )
+            self.w = np.asarray(coefs)
 
     def get_result(self):
         return {'beta': self.w.flatten()}