+++ benchopt/tests/test_benchmark_features.py
@@ -12,10 +12,7 @@ from benchopt.tests.utils import CaptureRunOutput
 
 
 def test_template_dataset():
-    datasets = {
-        "template_dataset.py": "raise ImportError()"
-    }
-
+    datasets = {"template_dataset.py": "raise ImportError()"}
     with temp_benchmark(datasets=datasets) as bench:
         # Make sure that importing template_dataset raises an error.
         with pytest.raises(ImportError):
@@ -122,10 +119,7 @@ def test_benchopt_min_version():
 @pytest.mark.parametrize('raise_install_error', [0, 1])
 def test_import_error_reporting(error, raise_install_error):
 
-    expected_exc = (
-        ImportError if raise_install_error and error is ImportError
-        else SystemExit
-    )
+    expected_exc = error if raise_install_error else SystemExit
 
     solver = """from benchopt import BaseSolver, safe_import_context
 
@@ -229,9 +223,8 @@ def test_objective_save_final_results(no_debug_log):
 
 def test_objective_cv_splitter(no_debug_log):
 
-    objective = """from benchopt import BaseObjective, safe_import_context
-        with safe_import_context() as import_ctx:
-            import numpy as np
+    objective = """from benchopt import BaseObjective
+        import numpy as np
 
         class Splitter():
             def split(self, X, y, groups=None):